requirements.txt
elasticsearch==8.15.0
click>=8.2.1
requests==2.31.0
pydantic>=2.0.0
jinja2==3.1.2
python-dotenv==1.0.0
jsonschema==4.20.0
sentence-transformers>=2.2.0
pyyaml==6.0.1
httpx==0.25.2
